requirements.txt
ansible-core<2.15.0
netaddr>=0.7.19
Jinja2>=2.11.3
treelib>=1.5.5
cvprac>=1.0.7
paramiko>=2.7.1
jsonschema>=4.5.1
requests>=2.25.1
PyYAML>=5.4.1
md-toc>=7.1.0
deepmerge>=1.0.1
cryptography>=38.0.4
pybatfish>=2022.9.7.77
rich>=13.3.1
fastapi>=0.95.0
